feat(tuci): ilan sponsorluğu ve canlı yayın blast TUCi ile ödeme
- AdCampaign: total_budget/spent_budget/cpc_bid Float → Integer (TUCi)
- Kampanya oluşturmada tuci_balance'dan peşin düşülür + TuciTransaction kaydı
- 402 dönülür yetersiz TUCi'de
- ad_service.py: Redis Lua script ve format string'ler integer'a geçti
- leads.py: COST_PER_PERSON 0.50 TL → 1 TUCi/kişi (tam sayı)
- wallet.py: spend_ad_campaign etiketi eklendi
- Flutter: boost dialog 100₺/1₺ → 100 TUCi/1 TUCi
- Flutter: blast dialog ₺ → TUCi
- analytics_service: estimatedCost double → int
- Migration: y0z1a2b3c4d5 Float→Integer için

+++ mobile/lib/services/analytics_service.dart
@@ -88,7 +88,7 @@ class AnalyticsService {
   static Future<Map<String, dynamic>?> sendLeadBlast({
     required String title,
     required String category,
-    required double estimatedCost,
+    required int estimatedCost,
     int? listingId,
   }) async {
     try {
